+++ src/Blammo/Logging/LoggingT.hs
@@ -4,7 +4,7 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
 
-module Blammo.Logging.LoggingT (LoggingT (..), runLoggerLoggingT) where
+module Blammo.Logging.LoggingT (LoggingT, runLoggerLoggingT) where
 
 import Prelude
 
@@ -30,7 +30,9 @@ import Control.Monad.Trans.Reader (ReaderT (..))
 import Control.Monad.Trans.Resource (MonadResource (..))
 import UnliftIO.Exception (finally)
 
-newtype LoggingT m a = LoggingT {runLoggingT :: Logger -> m a}
+type LoggingT = LoggingT' Logger
+
+newtype LoggingT' env m a = LoggingT {runLoggingT :: env -> m a}
   deriving
     ( Functor
     , Applicative
@@ -44,39 +46,40 @@ newtype LoggingT m a = LoggingT {runLoggingT :: Logger -> m a}
     , MonadMask
     , MonadResource
     )
-    via ReaderT Logger m
+    via ReaderT env m
   deriving
     ( MonadTrans
     )
-    via ReaderT Logger
+    via ReaderT env
 
-instance MonadBase b m => MonadBase b (LoggingT m) where
+instance MonadBase b m => MonadBase b (LoggingT' env m) where
   liftBase = lift . liftBase
 
-instance MonadTransControl LoggingT where
-  type StT LoggingT a = a
+instance MonadTransControl (LoggingT' env) where
+  type StT (LoggingT' env) a = a
   liftWith f = LoggingT $ \r -> f $ \(LoggingT t) -> t r
   restoreT = LoggingT . const
 
-instance MonadBaseControl b m => MonadBaseControl b (LoggingT m) where
-  type StM (LoggingT m) a = StM m a
+instance MonadBaseControl b m => MonadBaseControl b (LoggingT' env m) where
+  type StM (LoggingT' env m) a = StM m a
   liftBaseWith f = LoggingT $ \reader' ->
     liftBaseWith $ \runInBase ->
       f $ runInBase . (\(LoggingT r) -> r reader')
   restoreM = LoggingT . const . restoreM
 
-instance MonadIO m => MonadLogger (LoggingT m) where
-  monadLoggerLog a b c d = LoggingT $ \logger ->
+instance (MonadIO m, HasLogger env) => MonadLogger (LoggingT' env m) where
+  monadLoggerLog a b c d = do
+    logger <- getLogger
     liftIO $ runLogAction (loggerLogAction logger) a b c d
 
-instance MonadIO m => MonadLoggerIO (LoggingT m) where
-  askLoggerIO = LoggingT $ \logger ->
-    pure $ runLogAction $ loggerLogAction logger
+instance (MonadIO m, HasLogger env) => MonadLoggerIO (LoggingT' env m) where
+  askLoggerIO =
+    runLogAction . loggerLogAction <$> getLogger
 
-instance (Applicative m, Semigroup a) => Semigroup (LoggingT m a) where
+instance (Applicative m, Semigroup a) => Semigroup (LoggingT' env m a) where
   (<>) = liftA2 (<>)
 
-instance (Applicative m, Monoid a) => Monoid (LoggingT m a) where
+instance (Applicative m, Monoid a) => Monoid (LoggingT' env m a) where
   mempty = pure mempty
 
 runLoggerLoggingT
@@ -85,3 +88,6 @@ runLoggerLoggingT env f =
   runLoggingT f logger `finally` flushLogStr logger
  where
   logger = env ^. loggerL
+
+getLogger :: (Applicative m, HasLogger env) => LoggingT' env m Logger
+getLogger = LoggingT $ pure . (^. loggerL)