+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [v1.2.0.0](https://github.com/freckle/blammo/compare/v1.1.3.0...v1.2.0.0)
 
+- Defined new `LoggerT` instead of re-exporting from the `monad-logger` package
+
 ## [v1.1.3.0](https://github.com/freckle/blammo/compare/v1.1.2.3...v1.1.3.0)
 
 - Update fast-logger to fix log flushing bug, and remove 0.1s delay that was