+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## [v1.2.0.0](https://github.com/freckle/blammo/compare/v1.1.3.0...v1.2.0.0)
 
-- Defined new `LoggerT` instead of re-exporting from the `monad-logger` package
+- Defined new `LoggerT` instead of re-exporting from the `monad-logger` package.
+  This `LoggerT` is now internally a `ReaderT` of our `Logger` type rather than
+  of a `monad-logger` log action `Loc -> LogSource -> LogLevel -> LogStr -> IO ()`,
+  which makes it a bit more powerful as the latter can be obtained from the former.
 
 ## [v1.1.3.0](https://github.com/freckle/blammo/compare/v1.1.2.3...v1.1.3.0)