+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [v1.2.0.0](https://github.com/freckle/blammo/compare/v1.1.3.0...v1.2.0.0)
 
+- Add a new `withLogger` function that may serve as an alternative to
+  `newLogger` + `runLoggerLoggingT`.
 - Defined new `LoggerT` instead of re-exporting from the `monad-logger` package.
   This `LoggerT` is also a `ReaderT`-like newtype but based on our `Logger` type
   rather than rather than on a log action