+++ CHANGELOG.md
@@ -12,7 +12,9 @@
   `Logger`.
 - Added new `LoggerT'` which is a generalized variant of `LoggerT`; rather than
   being fixed to `Logger` as the reader context, it has a type parameter which
-  should be something with a `HasLogger` instance.
+  should be something with a `HasLogger` instance. Deriving via `LoggerT' app m`
+  can be a handy way to get an instance of `MonadLogger` for your application
+  monad.
 
 ## [v1.1.3.0](https://github.com/freckle/blammo/compare/v1.1.2.3...v1.1.3.0)